% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CI}
\alias{CI}
\title{Confidence Interval Data to compute AMIM}
\format{
## `CI`
A data datatable with the following columns:
\describe{
  \item{N}{Number of lags}
  \item{a}{Scale parameter equal to 1 as in Tran & Leivrik (2019)}
  \item{CI}{Confidence interval accordingly
     each number lags and scale parameter}
  ...
}
}
\source{
Tran & Leivrik (2019)
}
\usage{
CI
}
\description{
Confidence Interval Data to compute AMIM
}
\keyword{datasets}
