% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example Data to compute AMIM}
\format{
## `exampledata`
A data datatable with the following columns:
\describe{
  \item{Date}{Date format YYYY-MM-DD}
  \item{ticker}{Imaginary ticker}
  \item{RET}{Imaginary return}
  ...
}
}
\source{
Vu Le Tran
}
\usage{
exampledata
}
\description{
Example Data to compute AMIM
}
\keyword{datasets}
